% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_path.R
\name{movies_path}
\alias{movies_path}
\title{Movies path}
\usage{
movies_path(names, movies_dataset, movies_graph)
}
\arguments{
\item{names}{a vector filled with only two movie names}

\item{movies_dataset}{a dataframe with all the movies including a column with the titles}

\item{movies_graph}{a graph where vertices are movies and weighted edges between movies}
}
\value{
a dataframe with the shortest path between two movies
}
\description{
Movies path
}
\examples{
\dontrun{
movies_path(c("Superman", "Batman"), movies_dataset = tmdb, movies_graph = graph)
}

}
