% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_genre.R
\name{score_genre}
\alias{score_genre}
\title{score_genre}
\usage{
score_genre(i, j, genre_matrix)
}
\arguments{
\item{i}{the first movie}

\item{j}{the second movie}

\item{genre_matrix}{the matrix with all the genres and the movies}
}
\value{
an integer
}
\description{
score_genre
}
\examples{
\dontrun{
score_genre <- function(i,j, genre_matrix)
}
}
