% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_key.R
\name{score_key}
\alias{score_key}
\title{score_key}
\usage{
score_key(i, j, keywords_matrix)
}
\arguments{
\item{i}{the first movie}

\item{j}{the second movie}

\item{keywords_matrix}{the keyword matrix with the movies}
}
\value{
an integer
}
\description{
score_key
}
\examples{
\dontrun{
score_key(i,j, keywords_matrix = matrix_movies2)
}

}
