% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{similarities}
\alias{similarities}
\title{Similarities function}
\usage{
similarities(names, movies_dataset, sim_matrix)
}
\arguments{
\item{names}{a vector filled with two movies}

\item{movies_dataset}{a dataframe with all the movies (a column with the titles is used here called "original_title")}

\item{sim_matrix}{a matrix filled with the genres or keywords and the last column has the movie titles}
}
\value{
a list of genres in common
}
\description{
Similarities function
}
\examples{
\dontrun{
similarities_genre(c("The Dark Knight Rises","Superman Returns"),tmdb,matrix_movies)

}

}
