% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities_genre.R
\name{similarities_genre}
\alias{similarities_genre}
\title{similarities_genre}
\usage{
similarities_genre(names, movies_dataset, genre_matrix)
}
\arguments{
\item{names}{a vector filled with two movies}

\item{movies_dataset}{a dataframe with all the movies (a column with the titles is used here called "original_title")}

\item{genre_matrix}{a matrix filled with the genres and the last column has the movie titles}
}
\value{
a list of genres in common
}
\description{
similarities_genre
}
\examples{
\dontrun{
similarities_genre(c("The Dark Knight Rises","Superman Returns"),tmdb,matrix_movies)
}

}
