% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities_key.R
\name{similarities_key}
\alias{similarities_key}
\title{similarities_key}
\usage{
similarities_key(names, movies_dataset, key_matrix)
}
\arguments{
\item{names}{a vector filled with two movies}

\item{movies_dataset}{a dataframe with all the movies (a column with the titles is used here called "original_title")}

\item{key_matrix}{a matrix filled with the keywords and the last column has the movie titles}
}
\value{
a list
}
\description{
similarities_key
}
\examples{
\dontrun{
similarities_key(c("The Dark Knight Rises","Superman Returns"), tmdb, matrix_movies2)
}

}
