% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weight}
\alias{weight}
\title{weight function}
\usage{
weight(i, j, genre_matrix, key_matrix)
}
\arguments{
\item{i}{the first movie}

\item{j}{the second movie}

\item{genre_matrix}{a matrix with all the genres and movies}

\item{key_matrix}{a matrix with all the keywords and movies}
}
\value{
a weight
}
\description{
weight function
}
\examples{
\dontrun{
weight <-  function(i,j,genre_matrix = matrix_movies, key_matrix = matrix_movies2)
}

}
