% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_path.R
\name{movies_path}
\alias{movies_path}
\title{Movie path}
\usage{
movies_path(names, movies_dataset, movies_graph)
}
\arguments{
\item{names}{a vector filled with two movies}

\item{movies_dataset}{a dataframe with all the movies (a column with the titles is used here)}

\item{movies_graph}{a network showing all the edges between the nodes/movies}
}
\value{
a list
}
\description{
Movie path
}
\examples{
\dontrun{
movies_path(names, movies_dataset = tmdb, movies_graph = graph)
}

}
