% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_key.R
\name{score_key}
\alias{score_key}
\title{Score keywords}
\usage{
score_key(i, j, keywords_matrix)
}
\arguments{
\item{i}{integer, the first movie index}

\item{j}{integer, the second movie index}

\item{keywords_matrix}{matrix, rows correspond to movies and entries are the keywords}
}
\value{
an integer, the number of keywords in common
}
\description{
Score keywords
}
\examples{
\dontrun{
score_key(1,4, keywords_matrix = matrix_movies2)
}

}
