% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{similarities}
\alias{similarities}
\title{Similarities function}
\usage{
similarities(names, movies_dataset, sim_matrix)
}
\arguments{
\item{names}{a vector filled with only two movie names}

\item{movies_dataset}{a dataframe with all the movies including a column with the titles}

\item{sim_matrix}{a matrix with the genres or keywords and a column with the movie titles}
}
\value{
a list, of genres or keywords in common
}
\description{
Similarities function
}
\examples{
\dontrun{
similarities(c("The Dark Knight Rises","Superman Returns"),tmdb,matrix_movies)
similarities(c("The Dark Knight Rises","Superman Returns"),tmdb,matrix_movies2)
}

}
